# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LetsGoVideoAgent API",
    "summary": "Evidence-first general video understanding agent",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/videos/{video_id}": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "Get Video",
        "operationId": "get_video_api_v1_videos__video_id__get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/videos/imports": {
      "post": {
        "tags": [
          "videos"
        ],
        "summary": "Import Video",
        "operationId": "import_video_api_v1_videos_imports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/videos/uploads": {
      "post": {
        "tags": [
          "videos"
        ],
        "summary": "Upload Video",
        "operationId": "upload_video_api_v1_videos_uploads_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_video_api_v1_videos_uploads_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/videos/{video_id}/timeline": {
      "get": {
        "tags": [
          "timeline"
        ],
        "summary": "Get Timeline",
        "operationId": "get_timeline_api_v1_videos__video_id__timeline_get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/videos/{video_id}/questions": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Ask Video",
        "operationId": "ask_video_api_v1_videos__video_id__questions_post",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Video Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskQuestionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent-runs/{run_id}": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Agent Run",
        "operationId": "get_agent_run_api_v1_agent_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentRun": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          },
          "agent_version": {
            "type": "string",
            "title": "Agent Version"
          },
          "video_id": {
            "type": "string",
            "format": "uuid",
            "title": "Video Id"
          },
          "conversation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Conversation Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus",
            "default": "running"
          },
          "budget": {
            "$ref": "#/components/schemas/RunBudget"
          },
          "usage": {
            "$ref": "#/components/schemas/ModelUsage"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/AgentStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "stop_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Reason"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "agent_name",
          "agent_version",
          "video_id",
          "conversation_id",
          "budget"
        ],
        "title": "AgentRun",
        "description": "可展示的运行记录，不保存模型隐藏思维链。\n\nTrace 只记录节点、公开工具、输入摘要、结果摘要、用量和错误。这样既可排障，\n也不会把内部推理当作产品功能暴露给前端。"
      },
...
      "Answer": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          },
          "status": {
            "$ref": "#/components/schemas/AnswerStatus"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/EvidenceCitation"
            },
            "type": "array",
            "title": "Citations"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 0
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          },
          "trace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Trace Id"
          },
          "usage": {
            "$ref": "#/components/schemas/ModelUsage"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question_id",
          "status",
          "text",
          "trace_id"
        ],
        "title": "Answer"
      },
      "AnswerStatus": {
        "type": "string",
        "enum": [
          "answered",
          "partial",
          "abstained"
        ],
        "title": "AnswerStatus"
      },
      "AskQuestionRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          },
          "target": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/GlobalTarget"
              },
              {
                "$ref": "#/components/schemas/RangeTarget"
              },
              {
                "$ref": "#/components/schemas/MomentTarget"
              },
              {
                "$ref": "#/components/schemas/FrameTarget"
              }
            ],
            "title": "Target",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "frame": "#/components/schemas/FrameTarget",
                "global": "#/components/schemas/GlobalTarget",
                "moment": "#/components/schemas/MomentTarget",
                "range": "#/components/schemas/RangeTarget"
              }
            }
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "AskQuestionRequest"
      },
      "Body_upload_video_api_v1_videos_uploads_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "MP4/MOV/MKV/WebM 视频文件"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_video_api_v1_videos_uploads_post"
      },
...
      "GlobalTarget": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "global",
            "title": "Kind",
            "default": "global"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "GlobalTarget"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MomentTarget": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "moment",
            "title": "Kind",
            "default": "moment"
          },
          "timestamp_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Timestamp Ms"
          },
          "context_window_ms": {
            "type": "integer",
            "maximum": 60000.0,
            "minimum": 1000.0,
            "title": "Context Window Ms",
            "default": 8000
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "timestamp_ms"
        ],
        "title": "MomentTarget"
      },
...
      "RangeTarget": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "range",
            "title": "Kind",
            "default": "range"
          },
          "time_range": {
            "$ref": "#/components/schemas/TimeRange"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "time_range"
        ],
        "title": "RangeTarget"
      },
...
      "RunStatus": {
        "type": "string",
        "enum": [
          "running",
          "completed",
          "budget_exhausted",
          "policy_denied",
          "insufficient_evidence",
          "timed_out",
          "failed"
        ],
        "title": "RunStatus"
      },
...
      "SyntheticSource": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "synthetic",
            "title": "Kind",
            "default": "synthetic"
          },
          "fixture_name": {
            "type": "string",
            "title": "Fixture Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fixture_name"
        ],
        "title": "SyntheticSource"
      },
...
      "TimelineArtifact": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "video_id": {
            "type": "string",
            "format": "uuid",
            "title": "Video Id"
          },
          "kind": {
            "$ref": "#/components/schemas/TimelineKind"
          },
          "time_range": {
            "$ref": "#/components/schemas/TimeRange"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "speaker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 1
          },
          "observation_type": {
            "$ref": "#/components/schemas/ObservationType",
            "default": "direct"
          },
          "spatial_region": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpatialRegion"
              },
              {
                "type": "null"
              }
            ]
          },
          "snapshot_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Key"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "video_id",
          "kind",
          "time_range",
          "text",
          "provenance"
        ],
        "title": "TimelineArtifact",
        "description": "多轨时间轴上的最小知识单元。\n\n字幕、OCR、镜头、视觉描述可以在时间上重叠，因此这里不强迫它们合并为一棵\n唯一章节树。章节只是其中一条高层语义轨道。"
      },
...
      "TimelineResponse": {
        "properties": {
          "video_id": {
            "type": "string",
            "format": "uuid",
            "title": "Video Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TimelineArtifact"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "video_id",
          "items"
        ],
        "title": "TimelineResponse"
      },
      "UploadSource": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "upload",
            "title": "Kind",
            "default": "upload"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Size Bytes"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "original_filename",
          "content_type",
          "size_bytes"
        ],
        "title": "UploadSource"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Video": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Title"
          },
          "source": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/UploadSource"
              },
              {
                "$ref": "#/components/schemas/WebSource"
              },
              {
                "$ref": "#/components/schemas/SyntheticSource"
              }
            ],
            "title": "Source",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "synthetic": "#/components/schemas/SyntheticSource",
                "upload": "#/components/schemas/UploadSource",
                "web": "#/components/schemas/WebSource"
              }
            }
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "created"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "fps": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Fps"
          },
          "source_object_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Object Key"
          },
          "progress": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "default": 0
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version",
            "default": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "source"
        ],
        "title": "Video"
      },
...
      "WebImportRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 8,
            "title": "Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "rights_confirmed": {
            "type": "boolean",
            "title": "Rights Confirmed",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "url"
        ],
        "title": "WebImportRequest"
      },
      "WebSource": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "web",
            "title": "Kind",
            "default": "web"
          },
          "original_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Original Url"
          },
          "canonical_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Url"
          },
          "extractor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extractor"
          },
          "rights_confirmed": {
            "type": "boolean",
            "title": "Rights Confirmed",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "original_url"
        ],
        "title": "WebSource"
      }
    }
  }
}